Rename some files, added additional checks for more file types in CreateEDocument.php
