+++ app/Jobs/EDocument/CreateEDocument.php
@@ -1,0 +1,130 @@
+<?php
+/**
+ * Invoice Ninja (https://invoiceninja.com).
+ *
+ * @link https://github.com/invoiceninja/invoiceninja source repository
+ *
+ * @copyright Copyright (c) 2023. Invoice Ninja LLC (https://invoiceninja.com)
+ *
+ * @license https://www.elastic.co/licensing/elastic-license
+ */
+
+namespace App\Jobs\EDocument;
+
+use App\Models\Invoice;
+use App\Models\PurchaseOrder;
+use App\Models\Quote;
+use App\Services\EInvoicing\Standards\FacturaEInvoice;
+use App\Services\EInvoicing\Standards\ZugferdEDokument;
+use App\Utils\Ninja;
+use horstoeko\zugferd\ZugferdDocumentBuilder;
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\App;
+use function App\Jobs\Invoice\app;
+
+class CreateEDocument implements ShouldQueue
+{
+    use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
+
+    public $deleteWhenMissingModels = true;
+
+    public function __construct(private object $document, private bool $returnObject = false)
+    {
+    }
+
+    /**
+     * Execute the job.
+     *
+     * @return string|ZugferdDocumentBuilder
+     */
+    public function handle(): string|ZugferdDocumentBuilder
+    {
+        /* Forget the singleton*/
+        App::forgetInstance('translator');
+
+        /* Init a new copy of the translator*/
+        $t = app('translator');
+        /* Set the locale*/
+        App::setLocale($this->document->client->locale());
+
+        /* Set customized translations _NOW_ */
+        $t->replace(Ninja::transformTranslations($this->document->client->getMergedSettings()));
+
+        $e_document_type = $this->document->client->getSetting('e_invoice_type');
+        if ($this->document instanceof Invoice){
+            switch ($e_document_type) {
+                case "EN16931":
+                case "XInvoice_3_0":
+                case "XInvoice_2_3":
+                case "XInvoice_2_2":
+                case "XInvoice_2_1":
+                case "XInvoice_2_0":
+                case "XInvoice_1_0":
+                case "XInvoice-Extended":
+                case "XInvoice-BasicWL":
+                case "XInvoice-Basic":
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+
+                    return $this->returnObject ? $zugferd->xrechnung : $zugferd->getXml();
+                case "Facturae_3.2":
+                case "Facturae_3.2.1":
+                case "Facturae_3.2.2":
+                    return (new FacturaEInvoice($this->invoice, str_replace("Facturae_", "", $e_invoice_type)))->run();
+                default:
+
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+
+                    return $this->returnObject ? $zugferd : $zugferd->getXml();
+
+            }
+        }
+        elseif ($this->document instanceof Quote){
+            switch ($e_document_type){
+                case "EN16931":
+                case "XInvoice_3_0":
+                case "XInvoice_2_3":
+                case "XInvoice_2_2":
+                case "XInvoice_2_1":
+                case "XInvoice_2_0":
+                case "XInvoice_1_0":
+                case "XInvoice-Extended":
+                case "XInvoice-BasicWL":
+                case "XInvoice-Basic":
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+                    return $this->returnObject ? $zugferd->xrechnung : $zugferd->getXml();
+                default:
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+                    return $this->returnObject ? $zugferd : $zugferd->getXml();
+            }
+        }
+        elseif ($this->document instanceof PurchaseOrder){
+            switch ($e_document_type){
+                case "EN16931":
+                case "XInvoice_3_0":
+                case "XInvoice_2_3":
+                case "XInvoice_2_2":
+                case "XInvoice_2_1":
+                case "XInvoice_2_0":
+                case "XInvoice_1_0":
+                case "XInvoice-Extended":
+                case "XInvoice-BasicWL":
+                case "XInvoice-Basic":
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+                    return $this->returnObject ? $zugferd->xrechnung : $zugferd->getXml();
+                default:
+                    $zugferd = (new ZugferdEDokument($this->invoice))->run();
+                    return $this->returnObject ? $zugferd : $zugferd->getXml();
+            }
+        }
+        else{
+            return "";
+        }
+    }
+}
