+++ tests/Unit/EInvoiceTest.php
@@ -9,7 +9,7 @@
  * @license https://www.elastic.co/licensing/elastic-license
  */
 
-use App\Jobs\EInvoice\CreateEInvoice;
+use App\Jobs\EDocument\CreateEDocument;
 use App\Jobs\Entity\CreateRawPdf;
 use horstoeko\zugferd\ZugferdDocumentReader;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
@@ -41,7 +41,7 @@ class EInvoiceTest extends TestCase
         $this->company->e_invoice_type = "EN16931";
         $this->invoice->client->routing_id = 'DE123456789';
         $this->invoice->client->save();
-        $e_invoice = (new CreateEInvoice($this->invoice))->handle();
+        $e_invoice = (new CreateEDocument($this->invoice))->handle();
         $this->assertIsString($e_invoice);
     }
 
@@ -54,7 +54,7 @@ class EInvoiceTest extends TestCase
         $this->invoice->client->routing_id = 'DE123456789';
         $this->invoice->client->save();
 
-        $e_invoice = (new CreateEInvoice($this->invoice))->handle();
+        $e_invoice = (new CreateEDocument($this->invoice))->handle();
         $document = ZugferdDocumentReader::readAndGuessFromContent($e_invoice);
         $document->getDocumentInformation($documentno, $documenttypecode, $documentdate, $documentcurrency, $taxcurrency, $taxname, $documentlangeuage, $rest);
         $this->assertEquals($this->invoice->number, $documentno);
